+++ samples/sqltransport/multi-instance/SqlServer_2/Receiver/Program.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using NServiceBus;
-using NServiceBus.Transports.SQLServer;
 
 class Program
 {
@@ -28,20 +27,4 @@ class Program
         }
     }
 
-    static class ConnectionInfoProvider
-    {
-        public const string DefaultConnectionString = @"Data Source=.\SqlExpress;Database=ReceiverCatalog;Integrated Security=True";
-        const string SenderConnectionString = @"Data Source=.\SqlExpress;Database=SenderCatalog;Integrated Security=True";
-
-        public static ConnectionInfo GetConnection(string transportAddress)
-        {
-            var connectionString = transportAddress.StartsWith("Samples.SqlServer.MultiInstanceSender")
-                ? SenderConnectionString
-                : DefaultConnectionString;
-
-            return ConnectionInfo
-                    .Create()
-                    .UseConnectionString(connectionString);
-        }
-    }
 }
