+++ samples/sqltransport/multi-instance/SqlServer_2/Receiver/Receiver.csproj
@@ -38,11 +38,11 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ConnectionInfoProvider.cs" />
     <Compile Include="OrderHandler.cs" />
     <Compile Include="Program.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="app.config" />
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
