+++ samples/sqltransport/multi-instance/SqlServer_2/Sender/ConnectionInfoProvider.cs
@@ -1,0 +1,22 @@
+﻿using NServiceBus.Transports.SQLServer;
+
+#region SenderConnectionProvider
+
+static class ConnectionInfoProvider
+{
+    const string ReceiverConnectionString = @"Data Source=.\SqlExpress;Database=ReceiverCatalog;Integrated Security=True";
+    public const string DefaultConnectionString = @"Data Source=.\SqlExpress;Database=SenderCatalog;Integrated Security=True";
+
+    public static ConnectionInfo GetConnection(string transportAddress)
+    {
+        var connectionString = transportAddress.StartsWith("Samples.SqlServer.MultiInstanceReceiver")
+            ? ReceiverConnectionString
+            : DefaultConnectionString;
+
+        return ConnectionInfo
+                .Create()
+                .UseConnectionString(connectionString);
+    }
+}
+
+#endregion
