+++ samples/sqltransport/multi-instance/SqlServer_2/Sender/Program.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using Messages;
 using NServiceBus;
-using NServiceBus.Transports.SQLServer;
 
 public class Program
 {
@@ -48,24 +47,4 @@ public class Program
         }
     }
 
-    #region SenderConnectionProvider
-
-    static class ConnectionInfoProvider
-    {
-        const string ReceiverConnectionString = @"Data Source=.\SqlExpress;Database=ReceiverCatalog;Integrated Security=True";
-        public const string DefaultConnectionString = @"Data Source=.\SqlExpress;Database=SenderCatalog;Integrated Security=True";
-
-        public static ConnectionInfo GetConnection(string transportAddress)
-        {
-            var connectionString = transportAddress.StartsWith("Samples.SqlServer.MultiInstanceReceiver")
-                ? ReceiverConnectionString
-                : DefaultConnectionString;
-
-            return ConnectionInfo
-                    .Create()
-                    .UseConnectionString(connectionString);
-        }
-    }
-
-    #endregion
 }
