+++ samples/sqltransport/multi-instance/SqlServer_2/Sender/Sender.csproj
@@ -38,6 +38,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ConnectionInfoProvider.cs" />
     <Compile Include="OrderAcceptedHandler.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="ConfigErrorQueue.cs" />
