+++ samples/sqltransport/multi-instance/SqlServer_3/Receiver/OrderHandler.cs
@@ -1,16 +1,18 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Messages;
 using NServiceBus;
+using NServiceBus.Logging;
 
 public class OrderHandler :
     IHandleMessages<ClientOrder>
 {
+    static ILog log = LogManager.GetLogger<OrderHandler>();
+
     #region Reply
 
     public Task Handle(ClientOrder message, IMessageHandlerContext context)
     {
-        Console.WriteLine($"Handling ClientOrder with ID {message.OrderId}");
+        log.Info($"Handling ClientOrder with ID {message.OrderId}");
         var clientOrderAccepted = new ClientOrderAccepted
         {
             OrderId = message.OrderId
