+++ samples/sqltransport/multi-instance/SqlServer_3/Receiver/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Data.SqlClient;
 using System.Threading.Tasks;
 using NServiceBus;
 using NServiceBus.Transport.SQLServer;
@@ -39,22 +38,4 @@ class Program
             .ConfigureAwait(false);
     }
 
-    static class ConnectionProvider
-    {
-        const string DefaultConnectionString = @"Data Source=.\SqlExpress;Database=ReceiverCatalog;Integrated Security=True";
-        const string SenderConnectionString = @"Data Source=.\SqlExpress;Database=SenderCatalog;Integrated Security=True";
-
-        public static async Task<SqlConnection> GetConnection(string transportAddress)
-        {
-            var connectionString = transportAddress.StartsWith("Samples.SqlServer.MultiInstanceSender")
-                ? SenderConnectionString
-                : DefaultConnectionString;
-
-            var connection = new SqlConnection(connectionString);
-
-            await connection.OpenAsync().ConfigureAwait(false);
-
-            return connection;
-        }
-    }
 }
