+++ samples/sqltransport/multi-instance/SqlServer_3/Receiver/Receiver.csproj
@@ -39,6 +39,7 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ConnectionProvider.cs" />
     <Compile Include="OrderHandler.cs" />
     <Compile Include="Program.cs" />
   </ItemGroup>
