+++ samples/sqltransport/multi-instance/SqlServer_3/Sender/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Data.SqlClient;
 using System.Threading.Tasks;
 using Messages;
 using NServiceBus;
@@ -43,7 +42,7 @@ public class Program
                     return;
                 }
                 await PlaceOrder(endpointInstance)
-                .ConfigureAwait(false);
+                    .ConfigureAwait(false);
             }
         }
         finally
@@ -68,26 +67,4 @@ public class Program
 
         Console.WriteLine($"ClientOrder message sent with ID {order.OrderId}");
     }
-
-    #region SenderConnectionProvider
-    static class ConnectionProvider
-    {
-        const string DefaultConnectionString = @"Data Source=.\SqlExpress;Database=SenderCatalog;Integrated Security=True";
-        const string ReceiverConnectionString = @"Data Source=.\SqlExpress;Database=ReceiverCatalog;Integrated Security=True";
-
-        public static async Task<SqlConnection> GetConnection(string transportAddress)
-        {
-            var connectionString = transportAddress.StartsWith("Samples.SqlServer.MultiInstanceReceiver")
-                ? ReceiverConnectionString
-                : DefaultConnectionString;
-
-            var connection = new SqlConnection(connectionString);
-
-            await connection.OpenAsync().ConfigureAwait(false);
-
-            return connection;
-        }
-    }
-    #endregion
-
 }
